Prepare for supporting multiple modding platforms
