+++ common/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+	id "org.spongepowered.gradle.vanilla" version "0.2.1-SNAPSHOT"
+}
+
+minecraft {
+	version(rootProject.minecraft_version)
+}
