+++ common/src/main/java/nl/nielspoldervaart/gdmc/commands/SetBuildAreaCommand.java
