+++ common/src/main/java/nl/nielspoldervaart/gdmc/handlers/BuildAreaHandler.java
