+++ common/src/main/java/nl/nielspoldervaart/gdmc/handlers/StructureHandler.java
