+++ common/src/main/java/nl/nielspoldervaart/gdmc/utils/ChatComponentDataExtractor.java
