+++ common/src/main/resources/logo.png
