+++ gradle.properties
@@ -1,6 +1,6 @@
 # Sets default memory used for gradle commands. Can be overridden by user or command line properties.
 # This is required to provide enough memory for the Minecraft decompilation process.
-org.gradle.jvmargs=-Xmx3G
+org.gradle.jvmargs=-Xmx4096M
 org.gradle.daemon=true
 
 ## Mod Properties
@@ -13,7 +13,7 @@ mod_name=GDMC HTTP Interface
 # The license of the mod. Review your options at https://choosealicense.com/. All Rights Reserved is the default.
 mod_license=MIT
 # The mod version. See https://semver.org/
-mod_version=1.4.6
+mod_version=1.5.0
 # The group ID for the mod. It is only important when publishing as an artifact to a Maven repository.
 # This should match the base package used for the mod sources.
 # See https://maven.apache.org/guides/mini/guide-naming-conventions.html
@@ -27,4 +27,7 @@ mod_description=HTTP interface for reading and writing blocks (and more) with an
 # and what version is used automatically by build and run commands
 targetMinecraftVersion=1.20.2
 
+versionStr=
+
 # This build config setup is inspired by https://gitlab.com/jeseibel/distant-horizons
+# TODO integrate Fabric settings: https://github.com/Persson-dev/gdmc_http_fabric/blob/main/settings.gradle
