+++ versionProperties/1.19.2.properties
@@ -9,4 +9,4 @@ builds_for=forge
 # Forge
 forge_version=43.3.0
 forge_version_range=[43,)
-loader_version_range=[43,)
+forge_loader_version_range=[43,)
