+++ versionProperties/1.20.2.properties
@@ -9,4 +9,4 @@ builds_for=forge
 # Forge
 forge_version=48.1.0
 forge_version_range=[48,)
-loader_version_range=[48,)
+forge_loader_version_range=[48,)
